# [core] GL performance fixes