--- a/src/mbgl/programs/program_parameters.hpp
+++ b/src/mbgl/programs/program_parameters.hpp
@@ -6,16 +6,15 @@ namespace mbgl {
 
 class ProgramParameters {
 public:
-    ProgramParameters(float pixelRatio_ = 1.0,
-                      bool overdraw_ = false,
-                      const std::string& cacheDir_ = "")
-        : pixelRatio(pixelRatio_), overdraw(overdraw_), cacheDir(cacheDir_) {
-    }
+    ProgramParameters(float pixelRatio, bool overdraw, std::string cacheDir);
 
-    const float pixelRatio;
-    const bool overdraw;
+    const std::string defines;
+
+    std::string cachePath(const char* name) const;
+
+private:
+    const std::size_t hash;
     const std::string cacheDir;
 };
 
 } // namespace mbgl
-