--- a/src/mbgl/programs/program_parameters.hpp
+++ b/src/mbgl/programs/program_parameters.hpp
@@ -1,20 +1,24 @@
 #pragma once
 
+#include <mbgl/util/optional.hpp>
+
 #include <string>
+#include <vector>
 
 namespace mbgl {
 
 class ProgramParameters {
 public:
-    ProgramParameters(float pixelRatio, bool overdraw, std::string cacheDir);
+    ProgramParameters(float pixelRatio, bool overdraw, optional<std::string> cacheDir);
 
-    const std::string defines;
+    const std::string& getDefines() const;
+    optional<std::string> cachePath(const char* name) const;
 
-    std::string cachePath(const char* name) const;
+    ProgramParameters withAdditionalDefines(const std::vector<std::string>& defines) const;
 
 private:
-    const std::size_t hash;
-    const std::string cacheDir;
+    std::string defines;
+    optional<std::string> cacheDir;
 };
 
 } // namespace mbgl