… src/mbgl/renderer/painter_fill.cpp
… src/mbgl/renderer/painter_fill_extrusion.cpp
… src/mbgl/renderer/painter_symbol.cpp
… src/mbgl/shaders/circle.cpp
… src/mbgl/shaders/fill.cpp
… src/mbgl/shaders/fill_extrusion.cpp
… src/mbgl/shaders/fill_extrusion_pattern.cpp
… src/mbgl/shaders/fill_outline.cpp
… src/mbgl/shaders/fill_outline_pattern.cpp
… src/mbgl/shaders/fill_pattern.cpp
… src/mbgl/shaders/line.cpp
… src/mbgl/shaders/line_pattern.cpp
… src/mbgl/shaders/line_sdf.cpp
… src/mbgl/shaders/shaders.cpp
… src/mbgl/shaders/shaders.hpp
… src/mbgl/shaders/symbol_icon.cpp
… src/mbgl/shaders/symbol_sdf.cpp
… src/mbgl/style/layers/circle_layer_properties.hpp
… src/mbgl/style/layers/fill_extrusion_layer_properties.hpp
… src/mbgl/style/layers/fill_layer_properties.hpp
… src/mbgl/style/layers/line_layer_properties.hpp
… src/mbgl/style/layers/symbol_layer_properties.hpp
… src/mbgl/style/paint_property.hpp