Cherry-picks the following OpenGL performance and compatibility fixes to the release branch:

* #9251 -- Reduce number of varyings to 8 or less
* #9181 -- Store vertex attribute binding to prevent duplicate binds
* #9185 -- Dynamically compile shaders with uniforms instead of attributes for DDS
* #9257 -- Don't upload the FrameHistory texture in frames where it's not changing

```
2017-06-13 12:17:58.318728-0700 Bench GL[7283:1811481] | paris      | 60.0 fps |
2017-06-13 12:17:58.318824-0700 Bench GL[7283:1811481] | paris2     | 60.0 fps |
2017-06-13 12:17:58.318910-0700 Bench GL[7283:1811481] | alps       | 60.2 fps |
2017-06-13 12:17:58.318992-0700 Bench GL[7283:1811481] | us east    | 60.1 fps |
2017-06-13 12:17:58.319074-0700 Bench GL[7283:1811481] | greater la | 60.1 fps |
2017-06-13 12:17:58.319212-0700 Bench GL[7283:1811481] | sf         | 60.1 fps |
2017-06-13 12:17:58.319564-0700 Bench GL[7283:1811481] | oakland    | 60.2 fps |
2017-06-13 12:17:58.319654-0700 Bench GL[7283:1811481] | germany    | 59.9 fps |
2017-06-13 12:17:58.319726-0700 Bench GL[7283:1811481] Total FPS: 480.6
2017-06-13 12:17:58.319797-0700 Bench GL[7283:1811481] Average FPS: 60.1
```