@jfirebaugh 2017-06-13
Don't review yet; I'm going to repurpose this PR to include the other GL performance fixes like #9185.