@jfirebaugh 2017-06-13
For some reason this seems to be triggering timeouts on the Android Firebase tests. Can someone @mapbox/android take a look?