@zugaldia 2017-06-13
@jfirebaugh looks like Firebase was in a bad state, canceling the jobs and restarting them did the trick.