@jfirebaugh 2017-06-13
Thanks @zugaldia. This is now ready for review. @kkaefer could you post benchmark results from the iPad mini?