@kkaefer 2017-06-14
0c275ad6fa8d2c1e2523b6df7696d6fb09587411 (current master)

| name | Run 1 | Run 2 | Run 3 |
| --- | --- | --- | --- |
| `paris`      | 17.2 fps | 17.2 fps | 17.2 fps |
| `paris2`     | 19.4 fps | 18.5 fps | 19.3 fps |
| `alps`       |  8.8 fps |  8.9 fps |  8.8 fps |
| `us east`    | 24.4 fps | 25.6 fps | 25.0 fps |
| `greater la` | 17.6 fps | 18.0 fps | 18.0 fps |
| `sf`         | 22.5 fps | 21.7 fps | 21.6 fps |
| `oakland`    | 18.9 fps | 19.3 fps | 19.0 fps |
| `germany`    | 10.1 fps | 10.3 fps | 10.1 fps |


this PR

| name | Run 1 | Run 2 | Run 3 |
| --- | --- | --- | --- |
| `paris`      | 47.6 fps | 47.8 fps | 47.6 fps |
| `paris2`     | 59.8 fps | 59.9 fps | 59.9 fps |
| `alps`       | 22.2 fps | 22.2 fps | 22.2 fps |
| `us east`    | 59.9 fps | 59.8 fps | 59.9 fps |
| `greater la` | 47.4 fps | 47.5 fps | 47.5 fps |
| `sf`         | 60.1 fps | 60.3 fps | 60.3 fps |
| `oakland`    | 54.3 fps | 54.2 fps | 54.3 fps |
| `germany`    | 26.4 fps | 26.4 fps | 26.3 fps |